requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
opencv-python
Pillow

tensorflow
torch
torchvision

ultralytics

streamlit

jupyter
notebook
